chore: move file architecture
