+++ apps/web/src/query/query-key/types/query-key-type.ts
